bugfix: KernelPtr objects are owned by the user by default
Before this commit, KernelPtr objects created through its _from_ptr
constructor would not be owned by the user. This is incorrect,
in most cases the user is responsible for freeing up these pointers.

Only in the case of callbacks (e.g. for notifications and logging)
would the pointer not be owned by the user. However, this usage is
currently not implemented.

Fix this by making all KernelPtr objects owned by the user by default,
while still allowing this to be overridden in the future (e.g.
when notifications are more fully implemented) through the _from_ptr
constructor.