build: always install to _libs
Even when BITCOINKERNEL_LIB is provided, the binary
will always be copied to _libs/ for a more consistent
experience. This changes behaviour in that at runtime,
pbk will not inspect BITCOINKERNEL_LIB anymore, but
it will always use what is installed in _libs/.

Simplify CMakeLists.txt and make it more CMake-native.