chain: improve ChainstateManager.process_block behaviour
Introduce `ProcessBlockException` to throw when an invalid block
is processed.

Fix the non-functioning `new_block` out-parameter by changing it
to a return value.

Changes the process_block function signature by returning True/False
depending on whether the block was new, and raising a
ProcessBlockException for invalid blocks instead of returning it
as a return value.